requirements.txt
uvicorn[standard]
fastapi
pydantic
scikit-learn==1.6.1
pandas
joblib
httpx
setuptools>=78.1.1
supabase>=2.3.4
python-dotenv>=1.0.1